+++ src/structures/Core.js
@@ -2,17 +2,18 @@ const { Client } = require('discord.js');
 const path = require('path');
 const Store = require('./Store');
 const PermissionLevels = require('./PermissionLevels');
+const Collection = require('../util/Collection');
 const DB = require('./DB');
 const UniqueId = require('../util/UniqueId');
+const Config = require('./Config');
 
 const defaultOptions = {
-  typing: false,
   eventsFolder: 'events',
   commandsFolder: 'commands',
   token: null,
   prefix: undefined,
+  spaceAfterPrefix: false,
   splitArgs: ' ',
-  cmdsIn: ['text'],
   ignoreCase: true,
   permLevels: new PermissionLevels(),
   ignoreBots: true,
@@ -27,33 +28,31 @@ module.exports = class extends Client {
   constructor(options = {}) {
     options = { ...defaultOptions, ...options };
     const thisOptions = {
-      typing: options.typing,
       eventsFolder: options.eventsFolder,
       commandsFolder: options.commandsFolder,
       token: options.token,
       prefix: options.prefix,
+      spaceAfterPrefix: options.spaceAfterPrefix,
       splitArgs: options.splitArgs,
-      cmdsIn: options.cmdsIn,
       ignoreCase: options.ignoreCase,
       permLevels: options.permLevels,
       ignoreBots: options.ignoreBots,
       ignoreSelf: options.ignoreSelf,
       db: options.db,
     };
-    delete options.typing;
     delete options.eventsFolder;
     delete options.commandsFolder;
     delete options.token;
     delete options.prefix;
+    delete options.spaceAfterPrefix;
     delete options.splitArgs;
-    delete options.cmdsIn;
     delete options.ignoreCase;
     delete options.permLevels;
     delete options.ignoreBots;
     delete options.ignoreSelf;
     delete options.db;
     super(options);
-    const { cmdsIn, prefix, splitArgs, db } = thisOptions;
+    const { prefix, splitArgs, db } = thisOptions;
     if (
       db !== undefined &&
       db !== null &&
@@ -62,21 +61,15 @@ module.exports = class extends Client {
       const err = 'Db property must be instance of DB.';
       throw new Error(err);
     }
-    if (
-      (cmdsIn !== undefined &&
-        typeof cmdsIn !== 'string' &&
-        typeof cmdsIn !== 'object') ||
-      {}.hasOwnProperty.call(cmdsIn, 'join')
-    ) {
-      const err = 'CmdsIn option must be undefined, string or array.';
-      throw new TypeError(err);
-    }
-    if (cmdsIn === undefined) thisOptions.cmdsIn = [];
-    if (typeof cmdsIn === 'string') thisOptions.cmdsIn = [cmdsIn];
     if (prefix === undefined) thisOptions.prefix = '';
-    if (typeof prefix === 'object' && !{}.hasOwnProperty.call(prefix, 'test')) {
-      const err = 'Prefix option must be a string or regular expression.';
-      throw new TypeError(err);
+    if (
+      typeof prefix === 'object' &&
+      !(prefix instanceof RegExp) &&
+      !(prefix instanceof Array)
+    ) {
+      throw new TypeError(
+        'Prefix option must be a string or regular expression or array.'
+      );
     }
     if (
       splitArgs !== undefined &&
@@ -84,11 +77,11 @@ module.exports = class extends Client {
       splitArgs !== 'string' &&
       splitArgs !== 'object' &&
       typeof splitArgs === 'object' &&
-      !{}.hasOwnProperty.call(splitArgs, 'test')
+      typeof splitArgs.test !== 'function'
     ) {
-      const err =
-        'SplitArgs option must be a string, undefined, null or regular expression.';
-      throw new TypeError(err);
+      throw new TypeError(
+        'SplitArgs option must be a string, undefined, null or regular expression.'
+      );
     }
     this.public = {};
     /**
@@ -97,15 +90,16 @@ module.exports = class extends Client {
      * @private
      */
     this._private = {};
-    this._private.typing = thisOptions.typing;
+    this._private.sentPages = new Collection();
     this._private.eventsFolder = thisOptions.eventsFolder;
     this._private.commandsFolder = thisOptions.commandsFolder;
     this._private.fullpath = require.main.filename;
     this._private.filepath = path.basename(require.main.filename);
     this._private.dirpath = path.dirname(require.main.filename);
+    this.config = {};
+    this.config.guild = new Config(this, thisOptions);
     this.prefix = thisOptions.prefix;
     this.splitArgs = thisOptions.splitArgs;
-    this.cmdsIn = thisOptions.cmdsIn;
     this.ignoreCase = thisOptions.ignoreCase;
     this.permLevels = thisOptions.permLevels;
     this.ignoreBots = thisOptions.ignoreBots;