+++ src/util/Model.js
@@ -18,7 +18,7 @@ module.exports = class Model {
     this.Schema = new mongoose.Schema(options);
     this.Model = mongoose.model(this._modelName, this.Schema, this.name);
     this._db = this.Model.db;
-    this._toCollection();
+    this._db.on('connected', () => this._toCollection());
   }
 
   /**
@@ -26,11 +26,14 @@ module.exports = class Model {
    * @async
    */
   async getAll() {
-    const data = await this.Model.find({});
-    if (!data) return new Collection();
     const col = new Collection();
+    const data = await this.Model.find({});
+    if (!data) return col;
     for (const val of data) {
-      col.set(val._id, val._doc);
+      col.set(val._id.toHexString(), {
+        ...val._doc,
+        _id: val._id.toHexString(),
+      });
     }
     return col;
   }
@@ -64,6 +67,9 @@ module.exports = class Model {
    */
   findOne(query, value) {
     const data = this.collection.find(query, value);
+    if (data && data._id) {
+      data._id = new mongoose.mongo.ObjectID(data._id);
+    }
     return data ? { ...this.defaults, ...data } : this.defaults;
   }
 
@@ -71,17 +77,20 @@ module.exports = class Model {
    * @property {*} data
    * @returns {*} data
    * @example model.insertOne({ id: '1' });
+   * @async
    */
-  insertOne(data) {
+  async insertOne(data) {
     if (typeof data !== 'object') {
       const text = `First argument must be an object. Instead got ${typeof data}`;
       throw new TypeError(text);
     }
     data = { ...this.defaults, ...data };
     if (!data._id) data._id = new mongoose.mongo.ObjectID();
-    const col = this._db.collection(this.name);
-    this.collection.set(data._id, data);
-    col.insertOne(data);
+    this.collection.set(data._id.toHexString(), {
+      ...data,
+      _id: data._id.toHexString(),
+    });
+    await this._db.collection(this.name).insertOne(data);
     return data;
   }
 
@@ -90,8 +99,9 @@ module.exports = class Model {
    * @property {*} value
    * @returns {*} Deleted data.
    * @example model.deleteOne({ id: '1' });
+   * @async
    */
-  deleteOne(query, value) {
+  async deleteOne(query, value) {
     if (typeof query === 'string') {
       if (typeof value === 'undefined') {
         const text = 'Value must be specified.';
@@ -108,20 +118,9 @@ module.exports = class Model {
     if (!this.hasOne(query)) return null;
     const data = this.findOne(query);
     if (!data) return null;
-    this.collection.delete(data._id);
-    const col = this._db.collection(this.name);
-    col.deleteOne({ _id: data._id });
+    this.collection.delete(data._id.toHexString());
+    await this._db.collection(this.name).findOneAndDelete({ _id: data._id });
     return data;
-  }
-
-  async _updateOne(_id, value) {
-    const doc = await this.Model.findOne({ _id });
-    for (const key in value) {
-      if ({}.hasOwnProperty.call(value, key)) {
-        if (value[key] !== doc[key]) doc[key] = value[key];
-      }
-    }
-    doc.save();
   }
 
   /**
@@ -132,8 +131,9 @@ module.exports = class Model {
    * @example model.updateOne({ id: '1' }, { id: '2' });
    * @example model.updateOne(value => value.id === '1', { id: '2' });
    * @example model.updateOne('id', '1', { id: '2' });
+   * @async
    */
-  updateOne(query, value, newData) {
+  async updateOne(query, value, newData) {
     if (typeof query === 'string') {
       if (typeof value === 'undefined') {
         const text = 'Value must be specified.';
@@ -154,9 +154,15 @@ module.exports = class Model {
     }
     if (!this.hasOne(query)) return null;
     const data = this.findOne(query);
-    this.collection.set(data._id, value);
-    this._updateOne(data._id, value);
-    return value;
+    this.collection.set(data._id.toHexString(), {
+      ...data,
+      ...value,
+      _id: data._id.toHexString(),
+    });
+    await this._db
+      .collection(this.name)
+      .findOneAndUpdate({ _id: data._id }, { $set: value });
+    return data;
   }
 
   /**
@@ -167,9 +173,10 @@ module.exports = class Model {
    * @example model.upsertOne({ id: '1' }, { id: '2' });
    * @example model.upsertOne(value => value.id === '1', { id: '2' });
    * @example model.upsertOne('id', '1', { id: '2' });
+   * @async
    */
-  upsertOne(query, value, newData) {
-    const updated = this.updateOne(query, value, newData);
+  async upsertOne(query, value, newData) {
+    const updated = await this.updateOne(query, value, newData);
     if (updated) return updated;
     if (typeof query === 'function') query = {};
     if (typeof query === 'string') {
@@ -182,7 +189,7 @@ module.exports = class Model {
       query[prop] = value;
       value = newData;
     }
-    this.insertOne({ ...this.defaults, ...query, ...value });
+    await this.insertOne({ ...query, ...value });
     return { ...this.defaults, ...query, ...value };
   }
 };